+++ backend/liveness/liveness.go
@@ -9,8 +9,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/coreos/etcd/clientv3"
-	"github.com/coreos/etcd/mvcc/mvccpb"
+	"go.etcd.io/etcd/clientv3"
+	"go.etcd.io/etcd/mvcc/mvccpb"
 	"github.com/sensu/sensu-go/backend/store/etcd"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/time/rate"