+++ backend/store/etcd/compare.go
@@ -1,7 +1,7 @@
 package etcd
 
 import (
-	"github.com/coreos/etcd/clientv3"
+	"go.etcd.io/etcd/clientv3"
 	"github.com/sensu/sensu-go/types"
 )