+++ cli/client/testing/mock_cluster_client.go
@@ -1,6 +1,6 @@
 package testing
 
-import "github.com/coreos/etcd/clientv3"
+import "go.etcd.io/etcd/clientv3"
 
 // MemberList ...
 func (c *MockClient) MemberList() (*clientv3.MemberListResponse, error) {